# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse:
    properties:
      error:
        type: string
      fields: {}
    type: object
  github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse:
    properties:
      data: {}
    type: object
  internal_handler.ActOnApprovalRequest:
    properties:
      comment:
        type: string
      status:
        enum:
        - approved
        - rejected
        type: string
    required:
    - status
    type: object
...
  internal_handler.CreateApprovalSetupRequest:
    properties:
      department_id:
        type: string
      level_order:
        type: integer
      module_type:
        type: string
      role_id:
        type: string
    required:
    - level_order
    - module_type
    - role_id
    type: object
...
  internal_handler.SubmitForApprovalRequest:
    properties:
      department_id:
        type: string
      module_type:
        type: string
      reference_id:
        type: string
    required:
    - department_id
    - module_type
    - reference_id
    type: object
...
info:
  contact: {}
  description: REST API documentation for Workaround backend services.
  title: Workaround Backend API
  version: "1.0"
paths:
  /v1/approvals/pending:
    get:
      consumes:
      - application/json
      description: Returns pending approvals for the authenticated approver
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Get pending approvals
      tags:
      - Approvals
  /v1/approvals/reference/{reference_id}:
    get:
      consumes:
      - application/json
      description: Retrieves approvals for a module reference
      parameters:
      - description: Reference ID
        in: path
        name: reference_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Get approvals by reference
      tags:
      - Approvals
  /v1/approvals/reference/{reference_id}/act:
    post:
      consumes:
      - application/json
      description: Approves or rejects an approval request
      parameters:
      - description: Reference ID
        in: path
        name: reference_id
        required: true
        type: string
      - description: Approval action payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ActOnApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Act on approval
      tags:
      - Approvals
  /v1/approvals/setup:
    get:
      consumes:
      - application/json
      description: Retrieves approval chain by module type
      parameters:
      - description: Module type
        in: query
        name: module_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Get approval chain
      tags:
      - Approvals
    post:
      consumes:
      - application/json
      description: Creates an approval chain level for a module and department
      parameters:
      - description: Approval setup payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateApprovalSetupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Create approval setup
      tags:
      - Approvals
  /v1/approvals/setup/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an approval setup level by id
      parameters:
      - description: Approval setup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Delete approval setup
      tags:
      - Approvals
  /v1/approvals/submit:
    post:
      consumes:
      - application/json
      description: Submits a module record for approval
      parameters:
      - description: Submit for approval payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SubmitForApprovalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Submit for approval
      tags:
      - Approvals
...
schemes:
- http
- https
swagger: "2.0"